Correction in logic for 32-bit reservation flag

+++ src/scontrol/create_res.c
@@ -469,7 +469,7 @@ scontrol_create_res(int argc, char *argv[])
 	 */
 	if ((resv_msg.partition != NULL) && (resv_msg.node_list != NULL) &&
 	    (strcasecmp(resv_msg.node_list, "ALL") == 0)) {
-		if (resv_msg.flags == (uint16_t) NO_VAL)
+		if (resv_msg.flags == NO_VAL)
 			resv_msg.flags = RESERVE_FLAG_PART_NODES;
 		else
 			resv_msg.flags |= RESERVE_FLAG_PART_NODES;
@@ -481,7 +481,7 @@ scontrol_create_res(int argc, char *argv[])
 	 */
 	if ((resv_msg.partition == NULL) && (resv_msg.node_list != NULL) &&
 	    (strcasecmp(resv_msg.node_list, "ALL") == 0) &&
-	    (resv_msg.flags != (uint16_t) NO_VAL) &&
+	    (resv_msg.flags != NO_VAL) &&
 	    (resv_msg.flags & RESERVE_FLAG_PART_NODES)) {
 		exit_code = 1;
 		error("Part_Nodes flag requires specifying a Partition. "
